finalized/updated skeleton for pillars page
