+++ UCMERCED-SHPE-WEBSITE/src/Pillars/Pillars.css
@@ -1,0 +1,157 @@
+* {
+  border: 0;
+  padding: 0;
+  margin: 0;
+}
+
+.banner-container {
+  display: flex;
+  justify-content: center;
+  padding: 10px 15px;
+}
+
+.banner-image {
+  width: 1400px;
+  height: 310px;
+  object-fit: cover;
+  border-radius: 25px;
+}
+
+.mission-vision-container {
+  margin: 0rem 4rem;
+  text-align: center;
+}
+
+.mission-container {
+  display: inline-block;
+  border: 1px solid blue;
+  margin: 1rem 2rem;
+  padding: 1rem 1rem;
+  vertical-align: middle;
+  text-align: left;
+  max-width: 45%;
+  border-radius: 1rem;
+}
+
+.vision-container {
+  display: inline-block;
+  border: 1px solid blue;
+  margin: 1rem 2rem;
+  padding: 1rem 1rem;
+  vertical-align: middle;
+  text-align: left;
+  max-width: 45%;
+  border-radius: 1rem;
+}
+
+.vision-container p {
+  color: aqua;
+}
+
+.pillar-title {
+  display: flex;
+  margin: 20px;
+  margin-top: 30px;
+  background-color: rgb(198, 177, 198);
+  overflow: hidden;
+  height: 40vh;
+  box-shadow: 0px 12px 18px -6px rgba(0, 0, 0, 0.3);
+  border-radius: 10px 10px 10px 10px;
+  justify-content: center;
+}
+
+.pillar-container {
+  max-width: 80%;
+  margin: auto;
+}
+
+.fade-in-container {
+  margin: 0rem 2rem;
+  text-align: center;
+}
+
+.image-left-container {
+  display: inline-block;
+  margin: 2rem 1rem;
+  padding: 0rem 0rem;
+  vertical-align: middle;
+  text-align: left;
+  max-width: 45%;
+  border-radius: 1rem;
+}
+.text-right-container {
+  display: inline-block;
+  border: 1px solid blue;
+  margin: 2rem 2rem;
+  padding: 1rem 1rem;
+  vertical-align: middle;
+  text-align: left;
+  max-width: 42%;
+  border-radius: 1rem;
+}
+
+.pillar-image {
+  width: 500px;
+  height: 250px;
+  object-fit: cover;
+  border-radius: 25px;
+}
+
+.image-right-container {
+  display: inline-block;
+  margin: 2rem 1rem;
+  vertical-align: middle;
+  text-align: left;
+  max-width: 45%;
+  border-radius: 1rem;
+}
+.text-left-container {
+  display: inline-block;
+  border: 1px solid blue;
+  padding: 1rem 1rem;
+  margin: 0rem 0rem;
+  vertical-align: middle;
+  text-align: left;
+  max-width: 42%;
+  border-radius: 1rem;
+}
+
+/* 
+  .pillar-title {
+    display: flex;
+    align-items: center;
+    justify-content: left;
+    font-weight: bold;
+    width: 100%;
+    height: 100vh;
+  }
+  
+  .pillar-title:nth-child(1) {
+    background-image: url(../assets/SHPE-Familia.jpg);
+    background-size: cover;
+    background-attachment: fixed;
+  }
+  
+  .pillar-title:nth-child(2) {
+    background-image: url(../assets/SHPE-Retreat2021.png);
+    background-size: cover;
+    background-attachment: fixed;
+  }
+  
+  .pillar-title:nth-child(3) {
+    background-image: url(../assets/SHPE-Professionalism.jpg);
+    background-size: cover;
+    background-attachment: fixed;
+  }
+  
+  .pillar-title:nth-child(4) {
+    background-image: url(../assets/SHPE-Networking.png);
+    background-size: cover;
+    background-attachment: fixed;
+  }
+  
+  .pillar-title:nth-child(5) {
+    background-image: url(../assets/SHPE-Outreach1.png);
+    background-size: cover;
+    background-attachment: fixed;
+  } */
